refactor: move the remaining unrelated code out of lang
